Added Ecma-4a on suites, test.js doesnt exists anymore
